Removed NodesResources and moved things to NodesApiEndpoints Some changes connected to the movement of functions and classes
