+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/ManagementResources.scala
@@ -18,7 +18,9 @@ import pl.touk.nussknacker.engine.api.{Context, DisplayJson}
 import pl.touk.nussknacker.engine.testmode.TestProcess._
 import pl.touk.nussknacker.engine.util.json.BestEffortJsonEncoder
 import pl.touk.nussknacker.restmodel.{CustomActionRequest, CustomActionResponse}
-import pl.touk.nussknacker.ui.api.NodesResources.prepareTestFromParametersDecoder
+import pl.touk.nussknacker.ui.BadRequestError
+import pl.touk.nussknacker.ui.api.NodesApiEndpoints.Dtos.prepareTestFromParametersDecoder
+import pl.touk.nussknacker.ui.api.NodesApiEndpoints.TestFromParametersRequest
 import pl.touk.nussknacker.ui.api.ProcessesResources.ProcessUnmarshallingError
 import pl.touk.nussknacker.ui.metrics.TimeMeasuring.measureTime
 import pl.touk.nussknacker.ui.process.ProcessService
