+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/test/ScenarioTestService.scala
@@ -9,8 +9,9 @@ import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.definition.test.{TestInfoProvider, TestingCapabilities}
 import pl.touk.nussknacker.engine.testmode.TestProcess.TestResults
 import pl.touk.nussknacker.restmodel.definition.UISourceParameters
-import pl.touk.nussknacker.ui.api.{TestDataSettings, TestSourceParameters}
-import pl.touk.nussknacker.ui.definition.DefinitionsService
+import pl.touk.nussknacker.ui.api.NodesApiEndpoints.TestSourceParameters
+import pl.touk.nussknacker.ui.api.TestDataSettings
+import pl.touk.nussknacker.ui.definition.UIProcessObjectsFactory
 import pl.touk.nussknacker.ui.process.deployment.ScenarioTestExecutorService
 import pl.touk.nussknacker.ui.processreport.{NodeCount, ProcessCounter, RawCount}
 import pl.touk.nussknacker.ui.security.api.LoggedUser
@@ -42,7 +43,7 @@ class ScenarioTestService(
     val canonical = toCanonicalProcess(scenarioGraph, processName, isFragment)
     testInfoProvider
       .getTestParameters(canonical)
-      .map { case (id, params) => UISourceParameters(id, params.map(DefinitionsService.createUIParameter)) }
+      .map { case (id, params) => UISourceParameters(id, params.map(UIProcessObjectsFactory.createUIParameter)) }
       .toList
   }
