+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/validation/NodeValidator.scala
@@ -16,8 +16,8 @@ import pl.touk.nussknacker.engine.compile.nodecompilation.{
 import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
 import pl.touk.nussknacker.engine.variables.GlobalVariablesPreparer
 import pl.touk.nussknacker.restmodel.validation.PrettyValidationErrors
-import pl.touk.nussknacker.ui.api.{NodeValidationRequest, NodeValidationResult}
-import pl.touk.nussknacker.ui.definition.DefinitionsService
+import pl.touk.nussknacker.ui.api.NodesApiEndpoints.{NodeValidationRequest, NodeValidationResult}
+import pl.touk.nussknacker.ui.definition.UIProcessObjectsFactory
 import pl.touk.nussknacker.ui.process.fragment.FragmentRepository
 import pl.touk.nussknacker.ui.security.api.LoggedUser
 
@@ -37,7 +37,7 @@ class NodeValidator(modelData: ModelData, fragmentRepository: FragmentRepository
 
     // We create fragmentResolver for each request, because it requires LoggedUser to fetch fragments
     val fragmentResolver =
-      FragmentResolver(fragmentName => fragmentRepository.fetchLatestFragmentSync(fragmentName))
+      FragmentResolver(fragmentName => fragmentRepository.fetchLatestFragmentSync(fragmentName).map(_.canonical))
 
     nodeDataValidator.validate(
       nodeData.nodeData,
@@ -54,7 +54,7 @@ class NodeValidator(modelData: ModelData, fragmentRepository: FragmentRepository
           validationPerformed = false
         )
       case ValidationPerformed(errors, parameters, expressionType) =>
-        val uiParams = parameters.map(_.map(DefinitionsService.createUIParameter))
+        val uiParams = parameters.map(_.map(UIProcessObjectsFactory.createUIParameter))
 
         // We don't return MissingParameter error when we are returning those missing parameters to be added - since
         // it's not really exception ATM
